Extract Inner class and its Inner class
